metaschema: one file per concept
metaschema.go held five separable concerns; they now live in files
named for what they define: action.go (the edit-action vocabulary),
polarity.go (syntactic position), edit.go (the Edit type and its
enumeration), walker.go (the reflection walk deriving the edits), and
location.go (pattern matching, joined by MatchPattern from
non_contract.go, which is generic matching rather than non-contract
knowledge).
